requirements.txt
einops==0.6.1
matplotlib==3.7.2
numpy==1.22.4
PyYAML==6.0
tqdm==4.65.0
warmup_scheduler_pytorch==0.1.2
opencv-python==4.8.0.74
scipy==1.11.1
chumpy==0.70
wandb==0.15.8
PyOpenGL==3.1.0
trimesh==3.23.1
imageio==2.31.1
pyrender==0.1.45
h5py==3.8.0
cdflib==1.1.0
joblib==1.3.2
torchmetrics==1.1.1
omegaconf==2.3.0
scikit-learn==1.3.0
pycocotools==2.0.7
yacs==0.1.6
easydict==1.10
thop==0.1.1.post2209072238
torch_geometric==2.5.2
timm==0.9.16
pandas==2.2.2
pillow-heif==1.1.0
exifread==3.5.1
piexif==1.1.3